Revise German translations (de.json) (#35)
* Update spray translation (de.json)

"Spramotive" is missing the letter "y" ("Spraymotive"). In addition, the singular should be used instead of the plural.

* Update toy translation (de.json)

Epic Games uses the term "Spielsache" instead of "Spielzeug".

* Update wrap translation (de.json)

The singular should be used instead of the plural.

* Update translationNotAvailable transl. (de.json)

The previous translation sounds a bit clumsy and unprofessional.

* Update translationIncomplete translation (de.json)

Let's stay a little closer to the English phrase while making it sound nicer and more appropriate in this context. (also remove the period at the end of the sentence)

* Update nav translations (de.json)

- Although the words from the German translations start with a capital letter, we have to stick to the style of fnbr.co and use lower case. This affects the following translations: cosmetics [upcoming, list, gallery, icons, reminders, history],  about [dropdown, translate, about, privacy], account [logout]
- The correct German term for "privacy policy" is "Datenschutzerklärung". (about -> privacy)
- An appropriate and very common translation for "account" is "Konto". (account -> dropdown, account)
- An appropriate and very common translation for "logout" is "abmelden". (account -> logout)
- The previous translation sounds a bit clumsy. (currentLanguage)

* Update home translations (de.json)

Epic Games uses the term "Gegenstände" instead of "Items". Interestingly, they decided to change the translation for the item shop from "Gegenstand-Shop" to "Item Shop". It's best to stick to the translation of the game in order to improve usability.

* Update home->error translation (de.json)

The term "Abrufen" describes the context better than "einlesen". In addition, verb-derived nouns begin with an uppercase letter.

* Update home->error-too-long translation (de.json)

Let's stay a little closer to the English phrase while making it sound nicer and more appropriate in this context. (also remove the period at the end of the sentence)

* Update shop translations (de.json)

- Stick closer to the English phrase. (title)
- A couple of improvements and corrected mistakes in line1: Stick closer to the English phrase, add the missing "Fortnite" word and translate the tooltip.
- Remove the hyphen between "Item-Shop" to stay in sync with Epic Games translations. (line2)
- Separate infinitive group by adding a comma before "um".

* Update history translations (de.json)

Most of the updated translations don't follow the translation scheme and contain substantial orthographic and grammatical mistakes.

* Fix indentation for history (de.json)

* Update history->list->title translation (de.json)

Add missing hyphen between "Item Shop" and "Historie".

* Update news translations (de.json)

- Epic Games uses the term "Neuigkeiten" and not "Nachrichten" for news.
- The previous translations sound a bit clumsy and contains some orthographical and grammatical mistakes.
- Translate a phrase. (more-ingame-button)

* Update randomiser translations (de.json)

- Remove the hyphen and stick to the nav item. (title)
- Use "Gegenstand" instead of "Item". (line1)
- Improve translations for chooseText and randomise-all. I'm not quite happy with the last one, but I think it's still better than before.

* Update accounts translations (de.json)

Most of the previous translations don't follow the translation scheme, contain substantial orthographic and grammatical mistakes and some sound like Google Translate translations.

* Update errors translations (de.json)

Previous translations don't follow the translation scheme, contain orthographic and grammatical mistakes and sound a little clumsy.